+++ src/main/java/it/niedermann/fis/operation/remote/notification/sms/SmsProviderFactory.java
@@ -3,8 +3,9 @@ package it.niedermann.fis.operation.remote.notification.sms;
 import it.niedermann.fis.FisConfiguration;
 import it.niedermann.fis.main.model.OperationDto;
 import it.niedermann.fis.operation.remote.notification.OperationNotificationUtil;
-import it.niedermann.fis.operation.remote.notification.sms.sms77.Sms77Provider;
-import it.niedermann.fis.operation.remote.notification.sms.smsapi.SmsApiProvider;
+import it.niedermann.fis.operation.remote.notification.sms.com.smsapi.SmsApiProvider;
+import it.niedermann.fis.operation.remote.notification.sms.io.sms77.Sms77Provider;
+
 import org.apache.commons.lang3.NotImplementedException;
 import org.springframework.stereotype.Service;
